Rename useJoinToken to useJoinTokenSuspender
