+++ web/packages/teleport/src/Discover/Desktop/DesktopWrapper.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react';
 import { PingTeleportProvider } from 'teleport/Discover/Shared/PingTeleportContext';
 import { PING_INTERVAL } from 'teleport/Discover/Desktop/config';
 
-import { clearCachedJoinTokenResult } from 'teleport/Discover/Shared/useJoinToken';
+import { clearCachedJoinTokenResult } from 'teleport/Discover/Shared/useJoinTokenSuspender';
 
 import { ResourceKind } from '../Shared';
