+++ web/packages/teleport/src/Discover/Kubernetes/KubeWrapper.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 
 import { PingTeleportProvider } from 'teleport/Discover/Shared/PingTeleportContext';
 
-import { clearCachedJoinTokenResult } from 'teleport/Discover/Shared/useJoinToken';
+import { clearCachedJoinTokenResult } from 'teleport/Discover/Shared/useJoinTokenSuspender';
 
 import { ResourceKind } from '../Shared';
