+++ web/packages/teleport/src/Discover/Server/DownloadScript/DownloadScript.story.tsx
@@ -19,7 +19,7 @@ import { MemoryRouter } from 'react-router';
 
 import { Context as TeleportContext, ContextProvider } from 'teleport';
 import cfg from 'teleport/config';
-import { clearCachedJoinTokenResult } from 'teleport/Discover/Shared/useJoinToken';
+import { clearCachedJoinTokenResult } from 'teleport/Discover/Shared/useJoinTokenSuspender';
 import { PingTeleportProvider } from 'teleport/Discover/Shared/PingTeleportContext';
 import { userContext } from 'teleport/Main/fixtures';
 import { ResourceKind } from 'teleport/Discover/Shared';
