+++ web/packages/teleport/src/Discover/Server/ServerWrapper.tsx
@@ -18,7 +18,7 @@ import React, { useEffect } from 'react';
 
 import { PingTeleportProvider } from 'teleport/Discover/Shared/PingTeleportContext';
 
-import { clearCachedJoinTokenResult } from 'teleport/Discover/Shared/useJoinToken';
+import { clearCachedJoinTokenResult } from 'teleport/Discover/Shared/useJoinTokenSuspender';
 
 import { ResourceKind } from '../Shared';
